+++ src/lib/components/motion/AnimatePresenceStack.svelte
@@ -5,11 +5,11 @@
     $: mint = index + 1;
 </script>
 
-<Motion.div
+<div
     class="w-64 h-64 relative bg-gray-700/40 rounded-lg flex justify-center items-center"
 >
     <AnimatePresence initial={false} let:item list={[{ key: index }]}>
         <Card bind:index={mint} frontCard={false} />
         <Card bind:index drag="x" frontCard={true} />
     </AnimatePresence>
-</Motion.div>
+</div>