+++ src/routes/tests/+page.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
     import * as Card from "$lib/components/ui/card";
     import {
+        AnimateLayout,
         AnimatePresenceStack,
         AnimationSequence,
         ColorInterpolation,
@@ -372,6 +373,25 @@
                 <Card.Footer class="grid gap-1">
                     <h1 class="text-2xl font-semibold">Test Summary</h1>
                     <p class="text-green-500">Passed test</p>
+                </Card.Footer>
+            </Card.Footer>
+        </Card.Root>
+        <!-- End -->
+        <!-- Start -->
+        <Card.Root>
+            <Card.Header>
+                <Card.Title>Animate Layout</Card.Title>
+                <Card.Description>
+                    Click the switch to visualize the effect
+                </Card.Description>
+            </Card.Header>
+            <Card.Content>
+                <AnimateLayout />
+            </Card.Content>
+            <Card.Footer>
+                <Card.Footer class="grid gap-1">
+                    <h1 class="text-2xl font-semibold">Test Summary</h1>
+                    <p class="text-red-500">Failed test</p>
                 </Card.Footer>
             </Card.Footer>
         </Card.Root>